pyproject.toml
[tool.poetry]
name = "aulas"
version = "0.1.0"
description = ""
authors = ["AndersonStumpf <andersonreis.stumpf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.58.1"
jupyter = "^1.1.1"
python-dotenv = "^1.0.1"
yfinance = "^0.2.51"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
